// src/components/atoms/Button/types/types.ts
export enum ButtonType {
  button = 'button',
  reset = 'reset',
  submit = 'submit',
}

export enum ButtonSize {
  large = 'large',
  medium = 'medium',
  small = 'small',
}

export enum ButtonVariant {
  primary = 'primary',
  outline = 'outline',
  notification = 'notification',
  icon = 'icon',
}