// src/components/molecules/FormInput/types/types.ts
export enum InputType {
  text = 'text',
  email = 'email',
  password = 'password',
  file = 'file',
  textarea = 'textarea',
  select = 'select',
}

export enum InputId {
  login = 'login',
  password = 'password',
  passwordConfirm = 'passwordConfirm',
  nickname = 'nickname',
  captcha = 'captcha',
  gender = 'gender',
  files = 'files',
  messageText = 'messageText',
}

export enum InputSize {
  large = 'large',
  medium = 'medium',
}